task 5 edited

+++ task5/BitwiseOps.java
@@ -2,7 +2,7 @@ package com.hillel.task5;
 
 public class BitwiseOps {
     public static long bitDivByFour(double number) {
-        long result = (byte) number >> 2;
+        long result = (long) number >> 2;
         return result;
     }
 }
